Revert "refactor out util"
This reverts commit 6036c7a63382e0b99863cc58e1d88dd012d900fa.
