+++ package.json
@@ -11,15 +11,15 @@
     "test": "npm run transpile && npm run lasso:test && npm run clean && karma start --autoWatch false --singleRun true",
     "prepublish": "npm run clean && npm run lint && npm run test && npm run build",
     "build": "npm run transpile && npm run lasso",
-    "fix": "eslint src/index.js test/index.js --fix",
-    "lint": "eslint src/index.js > lint.txt",
+    "fix": "eslint src/index.js src/util.js test/index.js --fix",
+    "lint": "eslint src/index.js src/util.js > lint.txt",
     "lasso": "lasso require-run:./docs/index.js --out docs/static --inject-into docs/index.html --name bundle && npm run clean:lasso",
     "lasso:test": "lasso require-run:./test/index.js --name bundle-test --out test/static --config ./test/lasso-config.json && npm run clean:lasso",
     "clean": "rimraf reports .DS_Store test/.DS_Store src/.DS_Store docs/.DS_Store && npm run clean:lasso",
     "clean:lasso": "rimraf .cache build",
     "server": "browser-sync start -s --ss docs --index docs/index.html --files docs/index.html",
     "watch": "onchange src/*.js -- npm run build",
-    "transpile": "babel src/index.js --out-file index.js"
+    "transpile": "babel src/index.js --out-file index.js && babel src/util.js --out-file util.js"
   },
   "pre-commit": [
     "prepublish"
