+++ test/index.js
@@ -1,4 +1,5 @@
 var keyboardTrap = require('../index.js');
+var util = require('../util.js');
 var testData = require('./data.js');
 var trapEl;
 var trapNotified;
